requirements.txt
absl-py==2.1.0
aiosignal==1.3.2
altair==5.5.0
anyio==4.8.0
appnope==0.1.4
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
codecarbon==2.8.3
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
executing==2.2.0
fief-client==0.20.0
filelock==3.17.0
flatbuffers==25.2.10
flwr==1.15.2
fonttools==4.56.0
frozenlist==1.5.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.32.0
iterators==0.0.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.1.0
namex==0.0.8
narwhals==1.30.0
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-ml-py==12.570.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
questionary==2.1.0
RapidFuzz==3.12.1
ray==2.31.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.43.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==2.3.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tornado==6.4.2
traitlets==5.14.3
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
yaspin==3.1.0